🎨 Refacto modification actionnaire projector
