+++ packages/applications/projectors/src/index.ts
@@ -14,4 +14,4 @@ export * as PériodeProjector from './subscribers/période/période.projector';
 export * as ReprésentantLégalProjector from './subscribers/lauréat/représentant-légal';
 export * as HistoriqueProjector from './subscribers/historique/historique.projector';
 export * as ActionnaireProjector from './subscribers/lauréat/actionnaire.projector';
-export * as ModificationActionnaireProjector from './subscribers/lauréat/modificationActionnaire.projector';
+export * as ModificationActionnaireProjector from './subscribers/lauréat/modification-actionnaire';
