+++ packages/applications/projectors/src/subscribers/lauréat/modification-actionnaire/index.ts
@@ -1,0 +1,25 @@
+import { Message, MessageHandler, mediator } from 'mediateur';
+import { match } from 'ts-pattern';
+
+import { RebuildTriggered, Event } from '@potentiel-infrastructure/pg-event-sourcing';
+import { Actionnaire } from '@potentiel-domain/laureat';
+
+import { handleRebuilTriggered } from './handleRebuildTriggered';
+import { handleModificationActionnaireDemandée } from './handleModificationActionnaireDemandée';
+
+export type SubscriptionEvent = (Actionnaire.ActionnaireEvent & Event) | RebuildTriggered;
+
+export type Execute = Message<
+  'System.Projector.Lauréat.ModificationActionnaire',
+  SubscriptionEvent
+>;
+
+export const register = () => {
+  const handler: MessageHandler<Execute> = async (event) =>
+    match(event)
+      .with({ type: 'RebuildTriggered' }, handleRebuilTriggered)
+      .with({ type: 'ModificationActionnaireDemandée-V1' }, handleModificationActionnaireDemandée)
+      .otherwise(() => undefined);
+
+  mediator.register('System.Projector.Lauréat.ModificationActionnaire', handler);
+};
